Refactored multithread to use Thread Pool
